backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
deep-translator==1.11.4
python-dotenv==1.2.1
huggingface-hub==0.36.0
